Expicitely close the survey when done with it
This quiets an error that might have been due to the files being deleted before CoreData had released them.
The survey structs have gone out of score by the time the files are deleted, so the survyes and the core data classes should have ben released, but there is a lot in CoreData that is opaque, so I don't know.